Refactor image processing
