+++ react_native/App.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { View, Text, ScrollView, TouchableOpacity, StyleSheet, TextInput, Image, FlatList, Button } from 'react-native';
 import * as ImagePicker from 'expo-image-picker';
-import * as ImageManipulator from 'expo-image-manipulator';
+import { compressAndSquarePhoto } from './photoUploadUtils';
 
 // Mock AI label suggestions per inspection section
 const mockAISuggestions = {
@@ -54,40 +54,6 @@ const inspectionSections = [
   'Accessories & Conditions'
 ];
 
-// Resize then crop a photo to ensure a 1:1 aspect ratio
-async function compressAndSquarePhoto(uri) {
-  try {
-    const resized = await ImageManipulator.manipulateAsync(
-      uri,
-      [{ resize: { width: 1024 } }],
-      { compress: 0.7, format: ImageManipulator.SaveFormat.JPEG }
-    );
-
-    const { width, height } = resized;
-    if (width !== height) {
-      const size = Math.min(width, height);
-      const cropX = Math.floor((width - size) / 2);
-      const cropY = Math.floor((height - size) / 2);
-      const cropped = await ImageManipulator.manipulateAsync(
-        resized.uri,
-        [{
-          crop: {
-            originX: cropX,
-            originY: cropY,
-            width: size,
-            height: size,
-          },
-        }],
-        { compress: 0.7, format: ImageManipulator.SaveFormat.JPEG }
-      );
-      return cropped.uri;
-    }
-    return resized.uri;
-  } catch (err) {
-    console.error('Error squaring image:', err);
-    return uri;
-  }
-}
 
 export default function ClearSkyPhotoIntakeScreen() {
   const [photosBySection, setPhotosBySection] = useState({});
